LibMgmtSys.Backend/Application/Books/Queries/GetAllBooksQuery/GetAllBooksQueryHandler.cs
using LibMgmtSys.Backend.Application.Common.Interfaces.Persistence;
using LibMgmtSys.Backend.Domain.BookAggregate;
using MediatR;
using ErrorOr;

namespace LibMgmtSys.Backend.Application.Books.Queries.GetAllBooksQuery;

public class GetAllBooksQueryHandler : IRequestHandler<GetAllBooksQuery, ErrorOr<List<Book>>>
{
    private readonly IBookRepository _bookRepository;
    
    public GetAllBooksQueryHandler(IBookRepository bookRepository)
    {
        _bookRepository = bookRepository;
    }
    
    public async Task<ErrorOr<List<Book>>> Handle(GetAllBooksQuery request, CancellationToken cancellationToken)
    {
        return await _bookRepository.GetAllBooksAsync();
    }
}